% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{amod}
\alias{amod}
\title{Adjusted remainder function}
\usage{
amod(x, y)
}
\arguments{
\item{x}{an integer}

\item{y}{an integer}
}
\value{
an integer
}
\description{
Return the same as a \% b with b instead of 0.
}
\examples{
\dontrun{
amod(8, 4)
}
}
