% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armenian_calendar.R
\name{armenian_date}
\alias{armenian_date}
\title{Return the Armenian date data structure}
\usage{
armenian_date(year, month, day)
}
\arguments{
\item{year}{The Armeniam calendar year.}

\item{month}{The Armeniam calendar month.}

\item{day}{The Armeniam calendar day.}
}
\value{
An Armenian date.
}
\description{
Return the Armenian date data structure
}
\examples{
\dontrun{
armenian_date(1, 2, 3)
}
}
