% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armenian_calendar.R
\name{armenian_from_fixed}
\alias{armenian_from_fixed}
\title{Return the Armenian date corresponding to fixed date \code{date}}
\usage{
armenian_from_fixed(date)
}
\arguments{
\item{date}{A Rata Die.}
}
\value{
An Armenian calendar date.
}
\description{
Return the Armenian date corresponding to fixed date \code{date}
}
\examples{
\dontrun{
armenian_from_fixed(1234)
}
}
