% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{binary_search}
\alias{binary_search}
\title{Bisection search}
\usage{
binary_search(lo, hi, p, e)
}
\arguments{
\item{lo}{Lower end of the search interval.}

\item{hi}{Higher end of the search interval.}

\item{p}{A function (named, anonymous or formula).}

\item{e}{A function (named, anonymous or formula).}
}
\value{
A value inside the \verb{[lo, hi]} interval.
}
\description{
Bisection search for \code{x} in \verb{[lo, hi]} such that condition \code{e} holds,
\code{p} determines when to go left.
}
