% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcalr-package.R
\docType{package}
\name{calcalr-package}
\alias{calcalr}
\alias{calcalr-package}
\title{calcalr: Calendars of the World}
\description{
An R implementation of the algorithms described in the book Edward M. Reingold and Nachum Dershowitz "Calendrical Calculations: the Ultimate Edition".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://espinielli.github.io/calcalr/}
  \item \url{https://github.com/espinielli/calcalr}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@gmail.com} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

}
\keyword{internal}
