% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{clock_from_moment}
\alias{clock_from_moment}
\title{Return clock time hour:minute:second from moment 'tee'}
\usage{
clock_from_moment(tee)
}
\arguments{
\item{tee}{a double indicating a moment in time}
}
\value{
The clock time.
}
\description{
Return clock time hour:minute:second from moment 'tee'
}
\examples{
\dontrun{
clock_from_moment(10.2345)
}
}
