% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{degrees}
\alias{degrees}
\title{Convert  radians to degrees}
\usage{
degrees(rads)
}
\arguments{
\item{rads}{Radians.}
}
\value{
A number
}
\description{
Convert  radians to degrees
}
\examples{
\dontrun{
degrees(pi)
}
}
