% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egyptian_calendar.R
\name{egyptian_date}
\alias{egyptian_date}
\title{Return the Egyptian date data structure}
\usage{
egyptian_date(year, month, day)
}
\arguments{
\item{year}{The Egyptian calendar year.}

\item{month}{The Egyptian calendar month.}

\item{day}{The Egyptian calendar day.}
}
\value{
An Egyptian calendar date.
}
\description{
Return the Egyptian date data structure
}
\examples{
\dontrun{
egyptian_date(1, 2, 3)
}
}
