% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egyptian_calendar.R
\name{egyptian_from_fixed}
\alias{egyptian_from_fixed}
\title{Return the Egyptian date corresponding to fixed date \code{date}}
\usage{
egyptian_from_fixed(date)
}
\arguments{
\item{date}{A Rata Die.}
}
\value{
An Egyptian calendar date.
}
\description{
Return the Egyptian date corresponding to fixed date \code{date}
}
\examples{
\dontrun{
egyptian_from_fixed(1234)
}
}
