% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{epoch}
\alias{epoch}
\title{Epoch definition.}
\usage{
epoch()
}
\value{
a number indicating the start of the calendar
}
\description{
For Rata Diem, R.D., it is 0 (but any other reference would do.)
}
\examples{
\dontrun{
epoch()
}
}
