% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{final}
\alias{final}
\title{Return last integer greater or equal to initial index, i, such that condition, p, holds}
\usage{
final(i, p)
}
\arguments{
\item{i}{initial index}

\item{p}{A function (named, anonymous or formula).}
}
\value{
an integer
}
\description{
Return last integer greater or equal to initial index, i, such that condition, p, holds
}
\examples{
\dontrun{
final(0, \(i) i < 3)
}
}
