% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armenian_calendar.R
\name{fixed_from_armenian}
\alias{fixed_from_armenian}
\title{Return the fixed date corresponding to Armenian date \code{a_date}}
\usage{
fixed_from_armenian(a_date)
}
\arguments{
\item{a_date}{An Armenian calendar date.}
}
\value{
A Data Die.
}
\description{
Return the fixed date corresponding to Armenian date \code{a_date}
}
\examples{
\dontrun{
fixed_from_armenian(c(1, 2, 3))
}
}
