% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{fixed_from_jd}
\alias{fixed_from_jd}
\title{Return the fixed date corresponding to Julian day number \code{jd}}
\usage{
fixed_from_jd(jd)
}
\arguments{
\item{jd}{The Julian day.}
}
\value{
A fixed date.
}
\description{
Return the fixed date corresponding to Julian day number \code{jd}
}
\examples{
\dontrun{
fixed_from_jd(123.456)
}
}
