% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{fixed_from_mjd}
\alias{fixed_from_mjd}
\title{Return the fixed date corresponding to modified Julian day \code{mjd}}
\usage{
fixed_from_mjd(mjd)
}
\arguments{
\item{mjd}{The modified Julian day.}
}
\value{
A fixed date.
}
\description{
Return the fixed date corresponding to modified Julian day \code{mjd}
}
\examples{
\dontrun{
fixed_from_mjd(123.456)
}
}
