% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{fixed_from_moment}
\alias{fixed_from_moment}
\title{Return fixed date from moment}
\usage{
fixed_from_moment(tee)
}
\arguments{
\item{tee}{a double indicating a moment in time}
}
\value{
a Rata Die
}
\description{
Return fixed date from moment
}
\examples{
\dontrun{
fixed_from_moment(2.65)
}
}
