% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{hour}
\alias{hour}
\title{Return the hour of clock time \code{clock}}
\usage{
hour(clock)
}
\arguments{
\item{clock}{The clock time.}
}
\value{
The hour.
}
\description{
Return the hour of clock time \code{clock}
}
\examples{
\dontrun{
hour(3.5)
}
}
