% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{invert_angular}
\alias{invert_angular}
\title{Inverse of angular function}
\usage{
invert_angular(f, y, a, b, prec = 1e-05)
}
\arguments{
\item{f}{A function (named, anonymous or formula).}

\item{y}{A value where to invert \code{f}.}

\item{a}{Lower value of the search interval.}

\item{b}{Upper value of the search interval.}

\item{prec}{Required precision (default: 0.00001).}
}
\value{
A value.
}
\description{
Inverse of angular function
}
