% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{jd_from_fixed}
\alias{jd_from_fixed}
\title{Return the Julian day number corresponding to fixed date \code{rd}}
\usage{
jd_from_fixed(date)
}
\arguments{
\item{date}{a fixed date, i.e. a number of (eventually partial) days since epoch}
}
\value{
A Julian day.
}
\description{
Return the Julian day number corresponding to fixed date \code{rd}
}
\examples{
\dontrun{
jd_from_fixed(123.456)
}
}
