% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{jd_from_moment}
\alias{jd_from_moment}
\title{Return the Julian day number corresponding to moment \code{tee}}
\usage{
jd_from_moment(tee)
}
\arguments{
\item{tee}{a double indicating a moment in time}
}
\value{
The Julian day number.
}
\description{
Return the Julian day number corresponding to moment \code{tee}
}
\examples{
\dontrun{
jd_from_moment(123.456)
}
}
