% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles_of_days.R
\name{kday_before}
\alias{kday_before}
\title{Return the fixed date of the k-day before fixed date}
\usage{
kday_before(k, date)
}
\arguments{
\item{k}{the weekday; k=0 means Sunday, k=1 means Monday, and so on.}

\item{date}{a fixed date, i.e. a number of (eventually partial) days since epoch}
}
\value{
a fixed day number
}
\description{
Return the fixed date of the k-day before fixed date
}
\examples{
\dontrun{
kday_before(1, -61387)
}
}
\seealso{
Other cycles of days helpers: 
\code{\link{day_of_week_from_fixed}()},
\code{\link{kday_after}()},
\code{\link{kday_nearest}()},
\code{\link{kday_on_or_before}()}
}
\concept{cycles of days helpers}
