% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles_of_days.R
\name{kday_nearest}
\alias{kday_nearest}
\title{Return the fixed date of the k-day nearest fixed date 'date'}
\usage{
kday_nearest(k, date)
}
\arguments{
\item{k}{the weekday; k=0 means Sunday, k=1 means Monday, and so on.}

\item{date}{a fixed date, i.e. a number of (eventually partial) days since epoch}
}
\value{
a fixed day number
}
\description{
Return the fixed date of the k-day nearest fixed date 'date'
}
\examples{
\dontrun{
kday_nearest(1, -61387)
}
}
\seealso{
Other cycles of days helpers: 
\code{\link{kday_after}()},
\code{\link{kday_before}()},
\code{\link{kday_on_or_before}()}
}
\concept{cycles of days helpers}
