% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{minute}
\alias{minute}
\title{Return the minutes of clock time \code{clock}}
\usage{
minute(clock)
}
\arguments{
\item{clock}{The clock time.}
}
\value{
The minute
}
\description{
Return the minutes of clock time \code{clock}
}
\examples{
\dontrun{
minute(3.5)
}
}
