% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{mjd_from_fixed}
\alias{mjd_from_fixed}
\title{Return the modified Julian day corresponding to fixed date \code{rd}}
\usage{
mjd_from_fixed(date)
}
\arguments{
\item{date}{a fixed date, i.e. a number of (eventually partial) days since epoch}
}
\value{
The modified Julian day.
}
\description{
Return the modified Julian day corresponding to fixed date \code{rd}
}
\examples{
\dontrun{
mjd_from_fixed(123.456)
}
}
