% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mod3}
\alias{mod3}
\title{Interval modulus}
\usage{
mod3(x, a, b)
}
\arguments{
\item{x}{a real value}

\item{a}{the lower end of the interval}

\item{b}{the upper end of the interval}
}
\value{
an integer
}
\description{
An extension of the standard modulus which takes an interval as the modulus, rather than a divisor.
It shifts a real-valued \code{x} into the half-open real interval \verb{[a..b)}  by adding a multiple of
the length  \code{b - a}.
}
\examples{
\dontrun{
mod3(123, 7, 11)
}
}
