% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{moment_from_jd}
\alias{moment_from_jd}
\title{Return the moment corresponding to the Julian day number \code{jd}}
\usage{
moment_from_jd(jd)
}
\arguments{
\item{jd}{The Julian day.}
}
\value{
The moment representation of \code{jd}.
}
\description{
Return the moment corresponding to the Julian day number \code{jd}
}
\examples{
\dontrun{
moment_from_jd(123.456)
}
}
