% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nexti}
\alias{nexti}
\title{Return first integer greater or equal to initial index, i, such that condition, p, holds}
\usage{
nexti(i, p)
}
\arguments{
\item{i}{initial index}

\item{p}{A function (named, anonymous or formula).}
}
\value{
an integer
}
\description{
Return first integer greater or equal to initial index, i, such that condition, p, holds
}
\examples{
\dontrun{
nexti(0, \(i) i == 3)
}
}
