% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{quotient}
\alias{quotient}
\title{Return the whole part of m/n towards negative infinity}
\usage{
quotient(m, n)
}
\arguments{
\item{m}{The numerator.}

\item{n}{The denominator.}
}
\value{
An Integer.
}
\description{
Return the whole part of m/n towards negative infinity
}
\examples{
\dontrun{
quotient(4, 3)
}
}
