% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{radians}
\alias{radians}
\title{Convert degrees to radians}
\usage{
radians(degs)
}
\arguments{
\item{degs}{Degrees.}
}
\value{
A number
}
\description{
Convert degrees to radians
}
\examples{
\dontrun{
radians(30)
}
}
