% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{rd}
\alias{rd}
\title{Return rata diem (number of days since epoch) of moment in time, tee.}
\usage{
rd(tee)
}
\arguments{
\item{tee}{a double indicating a moment in time}
}
\value{
number of (eventually partial) days since epoch
}
\description{
Return rata diem (number of days since epoch) of moment in time, tee.
}
\examples{
\dontrun{
rd(1234.5)
}
}
