% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{seconds}
\alias{seconds}
\title{Return the seconds of clock time \code{clock}}
\usage{
seconds(clock)
}
\arguments{
\item{clock}{The clock time.}
}
\value{
The seconds
}
\description{
Return the seconds of clock time \code{clock}
}
\examples{
\dontrun{
seconds(3.5)
}
}
