% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{standard_day}
\alias{standard_day}
\title{Return the day of date \code{date}}
\usage{
standard_day(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The day component of a date.
}
\description{
Return the day of date \code{date}
}
\examples{
\dontrun{
standard_day(c(1, 2, 3))
}
}
