% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{standard_month}
\alias{standard_month}
\title{Return the month of date \code{date}}
\usage{
standard_month(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The month component of a date.
}
\description{
Return the month of date \code{date}
}
\examples{
\dontrun{
standard_month(c(1, 2, 3))
}
}
