% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{standard_year}
\alias{standard_year}
\title{Return the year of date \code{date}}
\usage{
standard_year(date)
}
\arguments{
\item{date}{A date.}
}
\value{
The year component of a date.
}
\description{
Return the year of date \code{date}
}
\examples{
\dontrun{
standard_year(c(1, 2, 3))
}
}
