% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summa}
\alias{summa}
\title{Return the sum of f(i) from i=k, k+1, ... till p(i) holds true or 0}
\usage{
summa(f, k, p)
}
\arguments{
\item{f}{A function (named, anonymous or formula).}

\item{k}{an integer index.}

\item{p}{A function (named, anonymous or formula).}
}
\value{
a number
}
\description{
This is a tail recursive implementation.
}
\examples{
\dontrun{
# sum of squares integers less than or equal to 4
summa(\(x) x**2, 1, \(i) i<=4)
}
}
