% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{time_from_clock}
\alias{time_from_clock}
\title{Return time of day from clock time \code{hms}}
\usage{
time_from_clock(hms)
}
\arguments{
\item{hms}{The time vector.}
}
\value{
The numerical representation of the clock time.
}
\description{
Return time of day from clock time \code{hms}
}
\examples{
\dontrun{
time_from_clock(c(11, 12, 13))
}
}
