% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_day_numbers.R
\name{time_of_day}
\alias{time_of_day}
\title{Return the time of day data structure}
\usage{
time_of_day(hour, minute, second)
}
\arguments{
\item{hour}{A number representing hours.}

\item{minute}{A number representing minutes.}

\item{second}{A number representing seconds.}
}
\value{
A representation of time of day, i.e. a vector with 3 numbers.
}
\description{
Return the time of day data structure
}
\examples{
\dontrun{
time_of_day(10, 25, 13)
}
}
