% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles_of_days.R
\name{day_of_week_from_fixed}
\alias{day_of_week_from_fixed}
\title{Return day of the week from a fixed date}
\usage{
day_of_week_from_fixed(date)
}
\arguments{
\item{date}{a fixed date, i.e. a number of (eventually partial) days since epoch}
}
\value{
an integer indicating the day of the week
}
\description{
Return day of the week from a fixed date
}
\examples{
\dontrun{
day_of_week_from_fixed(253427)
}
}
\seealso{
Other cycles of days: 
\code{\link{kday_after}()},
\code{\link{kday_before}()},
\code{\link{kday_nearest}()},
\code{\link{kday_on_or_before}()}
}
\concept{cycles of days}
